Move CompanyFormatToXml to Folder converters and use Enum
