+++ src/main/java/th/ac/kmitl/science/comsci/example/converters/CompanyFormatToXml.java
@@ -1,4 +1,6 @@
-package th.ac.kmitl.science.comsci.example.models;
+package th.ac.kmitl.science.comsci.example.converters;
+
+import th.ac.kmitl.science.comsci.example.models.Company;
