+++ src/test/java/th/ac/kmitl/science/comsci/example/models/CompanyFormatToXmlTest.java
@@ -1,6 +1,7 @@
 package th.ac.kmitl.science.comsci.example.models;
 
 import org.junit.Test;
+import th.ac.kmitl.science.comsci.example.converters.CompanyFormatToXml;
 import th.ac.kmitl.science.comsci.example.models.mocks.CompanyMock;
 
 public class CompanyFormatToXmlTest {
